fix: move constants.py to package root so cmd/main.py can import it
