+++ src/quartz_api/internal/service/v1/endpoint_types.py
@@ -15,7 +15,7 @@ from pydantic import (
 
 from quartz_api.internal import models
 
-from .constants import SUPPORT_EMAIL
+from quartz_api.constants import SUPPORT_EMAIL
 from .country_config import COUNTRIES
