+++ src/quartz_api/internal/service/v1/helpers.py
@@ -13,7 +13,7 @@ from quartz_api.internal import models
 from quartz_api.internal.middleware.auth import AuthDependency
 
 from .auth_scopes import ALL_COUNTRY_PERMISSIONS
-from .constants import SUPPORT_EMAIL
+from quartz_api.constants import SUPPORT_EMAIL
 from .country_config import (
     CountryConfig,
     RegionTypeConfig,
